+++ loritta-discord/src/main/java/net/perfectdreams/loritta/website/routes/api/v1/economy/PostTransferBalanceExternalRoute.kt
@@ -32,7 +32,7 @@ class PostTransferBalanceExternalRoute(loritta: LorittaDiscord) : RequiresAPIAut
 		val finalMoney = (garticos * transferRate)
 
 		transaction(Databases.loritta) {
-			profile.money += finalMoney
+			profile.money += finalMoney.toLong()
 
 			SonhosTransaction.insert {
 				it[givenBy] = null
